develop view order screen
