+++ app/src/main/java/com/udeni/e_commerce_system_development_task/ui/adaptor/OrderItemAdapter.java
@@ -1,113 +1,68 @@
 package com.udeni.e_commerce_system_development_task.ui.adaptor;
 
+import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.BaseAdapter;
 import android.widget.TextView;
-
-import androidx.recyclerview.widget.RecyclerView;
-
 import com.udeni.e_commerce_system_development_task.R;
-import com.udeni.e_commerce_system_development_task.data.local.model.Order;
+import com.udeni.e_commerce_system_development_task.data.local.model.OrderItem;
 
-import java.util.ArrayList;
 import java.util.List;
 
-public class OrderItemAdapter extends RecyclerView.Adapter<OrderItemAdapter.ViewHolder> {
+public class OrderItemAdapter extends BaseAdapter {
 
-    private List<Order> localDataSet;
-    private List<Order> orderListResult;
-    private OnItemClickListener listener;
+    private Context context;
+    private List<OrderItem> orderItems;
+    private LayoutInflater inflater;
 
-    public OrderItemAdapter(List<Order> dataSet) {
-        localDataSet = new ArrayList<>(dataSet);
-        orderListResult = dataSet;
+    public OrderItemAdapter(Context context, List<OrderItem> orderItems) {
+        this.context = context;
+        this.orderItems = orderItems;
+        this.inflater = LayoutInflater.from(context);
     }
 
     @Override
-    public ViewHolder onCreateViewHolder(ViewGroup viewGroup, int viewType) {
-        View view = LayoutInflater.from(viewGroup.getContext())
-                .inflate(R.layout.order_list_item, viewGroup, false);
-
-        return new ViewHolder(view);
+    public int getCount() {
+        return orderItems.size();
     }
 
     @Override
-    public void onBindViewHolder(ViewHolder viewHolder, final int position) {
-        Order order = orderListResult.get(position);
-        viewHolder.getTvOrderNumber().setText("Order #" + order.getOrderNumber());
-        viewHolder.getTvCustomerName().setText(order.getCustomer().getName());
-        viewHolder.getTvDateTime().setText(order.getDateTime());
-        viewHolder.getTvTotalAmount().setText(String.valueOf(order.calculateTotalAmount()));
-        viewHolder.itemView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (listener != null) {
-                    listener.onItemClick(order);
-                }
-            }
-        });
+    public Object getItem(int position) {
+        return orderItems.get(position);
     }
 
-    // Filter logic for search
-    public void filter(String text) {
-        orderListResult.clear();
-        if (text.isEmpty()) {
-            orderListResult.addAll(localDataSet);
+    @Override
+    public long getItemId(int position) {
+        return position;
+    }
+
+    @Override
+    public View getView(int position, View convertView, ViewGroup parent) {
+        ViewHolder holder;
+
+        if (convertView == null) {
+            convertView = inflater.inflate(R.layout.order_item, parent, false);
+            holder = new ViewHolder();
+            holder.itemName = convertView.findViewById(R.id.tv_item_name_qty);
+            holder.itemPrice = convertView.findViewById(R.id.tv_item_price);
+            convertView.setTag(holder);
         } else {
-            text = text.toLowerCase();
-            for (Order order : localDataSet) {
-                if (order.getOrderNumber().contains(text) ||
-                        order.getCustomer().getName().toLowerCase().contains(text) ||
-                        String.valueOf(order.getCustomer().getCustomerId()).contains(text)) {
-                    orderListResult.add(order);
-                }
-            }
+            holder = (ViewHolder) convertView.getTag();
         }
-        notifyDataSetChanged();
+
+        // Get the current order item
+        OrderItem orderItem = orderItems.get(position);
+
+        // Set the data to the views
+        holder.itemName.setText(orderItem.getItemCode().getName() + " X " + orderItem.getQty());
+        holder.itemPrice.setText(String.valueOf(orderItem.getItemPrice()));
+        return convertView;
     }
 
-    @Override
-    public int getItemCount() {
-        return orderListResult.size();
-    }
-
-    public void setOnItemClickListener(OnItemClickListener onItemClickListener) {
-        this.listener = onItemClickListener;
-    }
-
-    public interface OnItemClickListener {
-        void onItemClick(Order order);
-    }
-
-    public static class ViewHolder extends RecyclerView.ViewHolder {
-        private final TextView tvOrderNumber;
-        private final TextView tvCustomerName;
-        private final TextView tvDateTime;
-        private final TextView tvTotalAmount;
-
-        public ViewHolder(View view) {
-            super(view);
-            tvOrderNumber = view.findViewById(R.id.tvOrderNumber);
-            tvCustomerName = view.findViewById(R.id.tvCustomerName);
-            tvDateTime = view.findViewById(R.id.tvDateTime);
-            tvTotalAmount = view.findViewById(R.id.tvTotalAmount);
-        }
-
-        public TextView getTvOrderNumber() {
-            return tvOrderNumber;
-        }
-
-        public TextView getTvCustomerName() {
-            return tvCustomerName;
-        }
-
-        public TextView getTvDateTime() {
-            return tvDateTime;
-        }
-
-        public TextView getTvTotalAmount() {
-            return tvTotalAmount;
-        }
+    private static class ViewHolder {
+        TextView itemName;
+        TextView itemPrice;
     }
 }
