+++ app/src/main/res/values/dimens.xml
@@ -2,4 +2,11 @@
 <resources>
     <dimen name="list_item_height">50dp</dimen>
     <dimen name="margin_medium">20dp</dimen>
+    <dimen name="padding_small">8dp</dimen>
+    <dimen name="padding_medium">16dp</dimen>
+    <dimen name="text_size_large">36sp</dimen>
+    <dimen name="text_size_medium">18sp</dimen>
+    <dimen name="text_size_small">16sp</dimen>
+    <dimen name="divider_height">1dp</dimen>
+    <dimen name="margin_small">4dp</dimen>
 </resources>
