+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,12 @@
 <resources>
     <string name="app_name">E-commerce-System-Development-Task</string>
-    <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment1">Hello blank 1st fragment</string>
     <string name="hello_blank_fragment2">Hello blank 2nd fragment</string>
     <string name="element_text">order item</string>
-    <string name="order_number">Order #%s</string>
+    <string name="order_number">Order: #%s</string>
+    <string name="customer_name">Customer: %s</string>
+    <string name="date_time">Date time: %s</string>
+    <string name="lkr">LKR</string>
+    <string name="total">Total</string>
+    <string name="search_order">Search orders</string>
 </resources>
